Implemented functionality for the WaitList and incorporated
Incorporated functionality of waitlist into the CourseRoll.

Added a parameter to tell if the student is waitlisted.

Queue method makes a new copy of the list, dequeues, skips if matched,
then requeues the list from the copy.

Changed test methods to match the enrollment cap and altered the
waitlist functionality tests to pass.